backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "alexhcferreira",email = "alexcole1208@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "django-cpf-cnpj (>=1.0.0,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "faker (>=37.11.0,<38.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "pydotplus (>=2.0.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
